FIR completion: move logic computing CallableWeight to a more common place
Also rename it to `CallableMetadata` to avoid confusion.
